Change internal value of B field element from u128 to u64
This means that we should be able to do faster addition, as we can do
u64 addition instead of u128 addition. Note that it's not regular
addition but `overflowing_add` we have to use to achieve this type
change.

This change affects a lot of dependecies, as so many of our tests use B
field elements. This is why the diff for this commit is big.

This also removes one superfluous call to mod_reduce in the `Mul`
implementation for B field elements.

This speeds up NTT of 2^18 B field elements from 63.80ms to 13.98ms. So
it's a nearly factor 5 speedup!

It reduces the time for one Rescue Prime hash from 80µs to 55µs.

Both benchmarks were measured on my P1.

+++ benches/ntt_forward.rs
@@ -42,7 +42,7 @@ fn bfield_benchmark(
 
     let mut rng = rand::thread_rng();
     let mut xs = BFieldElement::random_elements(size, &mut rng);
-    let omega = xs[0].get_primitive_root_of_unity(size as u128).0.unwrap();
+    let omega = xs[0].get_primitive_root_of_unity(size as u64).0.unwrap();
 
     group.throughput(Throughput::Elements(size as u64));
     group.bench_with_input(bench_id, &size, |b, _| {
@@ -60,7 +60,7 @@ fn xfield_benchmark(
 
     let mut rng = rand::thread_rng();
     let mut xs = XFieldElement::random_elements(size, &mut rng);
-    let omega = xs[0].get_primitive_root_of_unity(size as u128).0.unwrap();
+    let omega = xs[0].get_primitive_root_of_unity(size as u64).0.unwrap();
 
     group.throughput(Throughput::Elements(size as u64));
     group.bench_with_input(bench_id, &size, |b, _| {
